tools-superlinear/rng_provider/rng_service_requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
rich>=13.0.0
pyserial>=3.5
pyusb>=1.2.1